Make scratches non obvious
