+++ tremor-pipeline/src/op/trickle/select.rs
@@ -14,6 +14,9 @@
 
 // [x] PERF0001: handle select without grouping or windows easier.
 
+#[cfg(test)]
+mod test;
+
 use super::window::{self, accumulate, get_or_create_group, Trait, Window, WindowEvent};
 use crate::{errors::Result, EventId, SignalKind};
 use crate::{op::prelude::*, EventIdGenerator};
@@ -40,7 +43,7 @@ pub struct TrickleSelect {
     pub event_id_gen: EventIdGenerator,
 }
 
-const NO_AGGRS: [InvokeAggrFn<'static>; 0] = [];
+pub(crate) const NO_AGGRS: [InvokeAggrFn<'static>; 0] = [];
 
 impl TrickleSelect {
     pub fn with_stmt(
@@ -169,7 +172,6 @@ impl Operator for TrickleSelect {
         let opts = Self::opts();
 
         let res = data.apply_select(select, |data, stmt| -> Result<Res> {
-
             let SelectStmt {
                 stmt: ref select,
                 ref mut aggregates,
@@ -179,9 +181,9 @@ impl Operator for TrickleSelect {
                 ref node_meta,
             } = stmt;
 
-            consts.window = Value::null();
-            consts.group = Value::null();
-            consts.args = Value::null();
+            consts.window = Value::const_null();
+            consts.group = Value::const_null();
+            consts.args = Value::const_null();
 
             let select: &Select = select;
 
@@ -221,9 +223,9 @@ impl Operator for TrickleSelect {
                 };
             }
 
-            let mut events = vec![];
+            let mut events = Vec::with_capacity(1);
 
-            let mut group_values =  if let Some(group_by) = &maybe_group_by {
+            let mut group_values = if let Some(group_by) = &maybe_group_by {
                 group_by.generate_groups(&ctx, data.value(), state, &node_meta, data.meta())?
             } else {
                 //
@@ -232,11 +234,7 @@ impl Operator for TrickleSelect {
                 // and we might drop it altogether based on having clause.
                 //
                 if windows.is_empty() {
-                    let group_value = Value::from(vec![()]);
-                    let group_str = sorted_serialize(&group_value)?;
-
-                    consts.group = group_value.clone_static();
-                    consts.group.push(group_str)?;
+                    consts.group = Value::from(vec![Value::const_null(), Value::from("[null]")]);
 
                     let (unwind_event, event_meta): (&mut Value, &mut Value) = data.parts_mut();
 
@@ -248,8 +246,7 @@ impl Operator for TrickleSelect {
                         recursion_limit: tremor_script::recursion_limit(),
                     };
                     let value =
-                        target
-                            .run(opts, &env, unwind_event, state, event_meta, &local_stack)?;
+                        target.run(opts, &env, unwind_event, state, event_meta, &local_stack)?;
 
                     let result = value.into_owned();
 
@@ -268,12 +265,13 @@ impl Operator for TrickleSelect {
                         } else {
                             Err(tremor_script::errors::query_guard_not_bool_err(
                                 select, guard, &test, &node_meta,
-                            ).into())
+                            )
+                            .into())
                         }
                     } else {
                         *unwind_event = result;
                         Ok(Res::Event)
-                    }
+                    };
                 };
                 vec![]
             };
@@ -285,11 +283,9 @@ impl Operator for TrickleSelect {
             // Handle eviction
             // retire the group data that didnt receive an event in `eviction_ns()` nanoseconds
             // if no event came after `2 * eviction_ns()` this group is finally cleared out
-            windows.iter_mut().for_each(|w|w.maybe_evict(ingest_ns));
+            windows.iter_mut().for_each(|w| w.maybe_evict(ingest_ns));
 
             let group_values: Vec<Value> = group_values.into_iter().map(Value::Array).collect();
-
-            let scratches = aggregate_scratches.as_mut();
 
             for group_value in group_values {
                 let group_str = sorted_serialize(&group_value)?;
@@ -351,8 +347,25 @@ impl Operator for TrickleSelect {
                             &group_str,
                             &group_value,
                         ));
-                        let window_event = stry!(this_group.window.on_event(&data, ingest_ns, origin_uri));
-                        if window_event.emit && !window_event.include {
+                        let window_event =
+                            stry!(this_group.window.on_event(&data, ingest_ns, origin_uri));
+
+                        // If we include the event we accumulate before we emit
+                        if window_event.include {
+                            stry!(accumulate(
+                                &ctx,
+                                consts.run(),
+                                opts,
+                                &node_meta,
+                                &local_stack,
+                                state,
+                                this_group,
+                                data,
+                                id,
+                                transactional
+                            ));
+                        }
+                        if window_event.emit {
                             // push
                             let env = Env {
                                 context: &ctx,
@@ -374,7 +387,7 @@ impl Operator for TrickleSelect {
                                 &data,
                                 ingest_ns,
                                 op_meta,
-                                    None,
+                                None,
                                 this_group.transactional
                             )) {
                                 events.push(port_and_event);
@@ -384,59 +397,20 @@ impl Operator for TrickleSelect {
                             this_group.reset();
                         }
 
-                        // accumulate
-
-                        // accumulate the current event
-                        let env = Env {
-                            context: &ctx,
-                            consts: consts.run(),
-                            aggrs: &NO_AGGRS,
-                            meta: &node_meta,
-                            recursion_limit: tremor_script::recursion_limit(),
-                        };
-                        stry!(accumulate(
-                            opts,
-                            &node_meta,
-                            &env,
-                            &local_stack,
-                            state,
-                            this_group,
-                            data,
-                            id,
-                            transactional
-                        ));
-
-                        if window_event.emit && window_event.include {
-                            // push
-                            let env = Env {
-                                context: &ctx,
-                                consts: consts.run(),
-                                aggrs: &this_group.aggrs,
-                                meta: &node_meta,
-                                recursion_limit: tremor_script::recursion_limit(),
-                            };
-
-                            let mut outgoing_event_id = event_id_gen.next_id();
-                            mem::swap(&mut outgoing_event_id, &mut this_group.id);
-                            if let Some(port_and_event) = stry!(execute_select_and_having(
-                                &select,
-                                &node_meta,
+                        // If we do not include the event we accumulate after we emit
+                        if !window_event.include {
+                            stry!(accumulate(
+                                &ctx,
+                                consts.run(),
                                 opts,
+                                &node_meta,
                                 &local_stack,
                                 state,
-                                &env,
-                                outgoing_event_id,
-                                &data,
-                                ingest_ns,
-                                op_meta,
-                                    None,
-                                this_group.transactional
-                            )) {
-                                events.push(port_and_event);
-                            };
-                            // re-initialize aggr state for new window
-                            // reset transactional state for outgoing events
-                            this_group.reset();
+                                this_group,
+                                data,
+                                id,
+                                transactional
+                            ));
                         }
                     }
                     _ => {
@@ -447,21 +421,21 @@ impl Operator for TrickleSelect {
                         // }
                         // for window in emit_windows {
                         //   if include
-                        //      this.aggrs.merge(scratch);
+                        //      this.aggrs.merge(scratch_last_window);
                         //      push event
-                        //      swap(scratch, aggrs)
+                        //      swap(scratch_last_window, aggrs)
                         //      aggrs.init() // init next window
                         //   else
                         //      push event
                         //      if first {
-                        //          swap(scratch, aggrs)
-                        //          aggrs.init() // no merge needed
+                        //          swap(scratch_last_window, aggrs)
+                        //          this.aggrs.init() // no merge needed
                         //      } else {
-                        //          local_scratch = aggregates.clone()
-                        //          swap(scratch, local_scratch) // store old aggrs state from last window
-                        //          swap(aggrs, scratch) // store aggrs state from this before init
-                        //          aggrs.init()
-                        //          this.aggrs.merge(local_scratch);
+                        //          scratch2 = this.aggrs.clone()
+                        //          swap(scratch_last_window, scratch2) // store old aggrs state from last window
+                        //          swap(this.aggrs, scratch_last_window) // store aggrs state from this before init
+                        //          this.aggrs.init()
+                        //          this.aggrs.merge(scratch2);
                         //      }
                         // }
                         // if emit && !include { // if the first window is !include we postpone accumulation until here
@@ -482,7 +456,7 @@ impl Operator for TrickleSelect {
                         // we need to store the state before re-init in a scratch, so the next window can pull it whenever
                         // it is ready to do so, before or after accumulating
                         //
-                        // event | size 2 | tilt | size 3 |
+                        // event | size 2 | tilt   | size 3 |
                         //     1 | [1]    |        |
                         //     2 | [1, 2] |        |
                         //     3 | emit   | [1, 2] | [1, 2]
@@ -499,46 +473,238 @@ impl Operator for TrickleSelect {
                         //       |        |        | emit                     // the next window we collect one too many elements
 
                         // we need scratches for multiple window handling, to avoid allocating on the hot path here
-                        if let Some((scratch1, scratch2)) = scratches {
-                            // scratches for passing on the event id between windows to track all events contributing to an outgoing windowed event
-                            // cannot put them in AggregateScratch, as this is in tremor_script and doesnt know about EventId
-                            // cannot put AggregateScratch in TrickleSelect as it has lifetimes which we tried to avoid here
-                            let mut event_id_scratch1 = EventId::default();
-                            let mut event_id_scratch2 = EventId::default();
+                        let (scratch_last_window, scratch2) = aggregate_scratches;
+                        // scratches for passing on the event id between windows to track all events contributing to an outgoing windowed event
+                        // cannot put them in AggregateScratch, as this is in tremor_script and doesnt know about EventId
+                        // cannot put AggregateScratch in TrickleSelect as it has lifetimes which we tried to avoid here
+                        let mut event_id_scratch_last_window = EventId::default();
+                        let mut event_id_scratch2 = EventId::default();
 
-                            // gather window events
-                            let mut emit_window_events = Vec::with_capacity(windows.len());
-                            let step1_iter = windows.iter_mut();
-                            for window in step1_iter {
-                                consts.window = Value::from(window.name().to_string());
-                                // get current window group
-                                let this_group = stry!(get_or_create_group(
-                                    window,
-                                    event_id_gen,
-                                    aggregates,
-                                    &group_str,
-                                    &group_value,
-                                ));
-                                let window_event = stry!(this_group.window.on_event(&data, ingest_ns, origin_uri));
-                                if window_event.emit {
-                                    emit_window_events.push(window_event);
+                        // gather window events
+                        let mut emit_window_events = Vec::with_capacity(windows.len());
+                        let step1_iter = windows.iter_mut();
+                        for window in step1_iter {
+                            consts.window = Value::from(window.name().to_string());
+                            // get current window group
+                            let this_group = stry!(get_or_create_group(
+                                window,
+                                event_id_gen,
+                                aggregates,
+                                &group_str,
+                                &group_value,
+                            ));
+                            let window_event =
+                                stry!(this_group.window.on_event(&data, ingest_ns, origin_uri));
+                            if window_event.emit {
+                                emit_window_events.push(window_event);
+                            } else {
+                                break;
+                            }
+                        }
+
+                        // accumulate if we do not emit anything or the first emitting window includes the current event into its state (window_event.include == true)
+                        let pending_accumulate = if let Some(WindowEvent {
+                            emit: true,
+                            include: false,
+                            ..
+                        }) = emit_window_events.first()
+                        {
+                            // postpone
+                            true
+                        } else if let Some(first_window) = windows.first_mut() {
+                            consts.window = Value::from(first_window.name().to_string());
+
+                            // get current window group
+                            let this_group = stry!(get_or_create_group(
+                                first_window,
+                                event_id_gen,
+                                aggregates,
+                                &group_str,
+                                &group_value,
+                            ));
+
+                            // accumulate the current event
+                            stry!(accumulate(
+                                &ctx,
+                                consts.run(),
+                                opts,
+                                &node_meta,
+                                &local_stack,
+                                state,
+                                this_group,
+                                data,
+                                id,
+                                transactional
+                            ));
+                            false
+                        } else {
+                            // should not happen
+                            false
+                        };
+
+                        // merge previous aggr state, push window event out, re-initialize for new window
+                        let mut first = true;
+                        let emit_window_iter = emit_window_events.iter().zip(windows.iter_mut());
+                        for (window_event, window) in emit_window_iter {
+                            consts.window = Value::from(window.name().to_string());
+
+                            // get current window group
+                            let this_group = stry!(get_or_create_group(
+                                window,
+                                event_id_gen,
+                                aggregates,
+                                &group_str,
+                                &group_value,
+                            ));
+
+                            if window_event.include {
+                                // add this event to the aggr state **BEFORE** emit and propagate to next windows
+                                // merge with scratch
+                                if !first {
+                                    for (this, prev) in this_group
+                                        .aggrs
+                                        .iter_mut()
+                                        .zip(scratch_last_window.aggregates.iter())
+                                    {
+                                        this.invocable.merge(&prev.invocable).map_err(|e| {
+                                            let r: Option<&Registry> = None;
+                                            e.into_err(prev, prev, r, &node_meta)
+                                        })?;
+                                    }
+                                    this_group
+                                        .track_transactional(scratch_last_window.transactional);
+                                    this_group.id.track(&event_id_scratch_last_window);
+                                }
+                                // push event
+                                let env = Env {
+                                    context: &ctx,
+                                    consts: consts.run(),
+                                    aggrs: &this_group.aggrs,
+                                    meta: &node_meta,
+                                    recursion_limit: tremor_script::recursion_limit(),
+                                };
+
+                                let mut outgoing_event_id = event_id_gen.next_id();
+                                event_id_scratch_last_window = this_group.id.clone(); // store the id for to be tracked by the next window
+                                mem::swap(&mut outgoing_event_id, &mut this_group.id);
+                                if let Some(port_and_event) = stry!(execute_select_and_having(
+                                    &select,
+                                    &node_meta,
+                                    opts,
+                                    &local_stack,
+                                    state,
+                                    &env,
+                                    outgoing_event_id,
+                                    &data,
+                                    ingest_ns,
+                                    op_meta,
+                                    None,
+                                    this_group.transactional
+                                )) {
+                                    events.push(port_and_event);
+                                };
+
+                                // swap(aggrs, scratch) - store state before init to scratch as we need it for the next window
+                                std::mem::swap(
+                                    &mut scratch_last_window.aggregates,
+                                    &mut this_group.aggrs,
+                                );
+                                // store transactional state for mixing into the next window
+                                scratch_last_window.transactional = this_group.transactional;
+
+                                // aggrs.init() and reset transactional state
+                                this_group.reset();
+                            } else {
+                                // add this event to the aggr state **AFTER** emit and propagate to next windows
+
+                                // push event
+                                let env = Env {
+                                    context: &ctx,
+                                    consts: consts.run(),
+                                    aggrs: &this_group.aggrs,
+                                    meta: &node_meta,
+                                    recursion_limit: tremor_script::recursion_limit(),
+                                };
+
+                                let mut outgoing_event_id = event_id_gen.next_id();
+                                if !first {
+                                    // store old event_ids state from last window into scratch2
+                                    std::mem::swap(
+                                        &mut event_id_scratch_last_window,
+                                        &mut event_id_scratch2,
+                                    );
+                                }
+                                event_id_scratch_last_window = this_group.id.clone(); // store the id for to be tracked by the next window
+                                mem::swap(&mut outgoing_event_id, &mut this_group.id);
+                                if let Some(port_and_event) = stry!(execute_select_and_having(
+                                    &select,
+                                    &node_meta,
+                                    opts,
+                                    &local_stack,
+                                    state,
+                                    &env,
+                                    outgoing_event_id,
+                                    &data,
+                                    ingest_ns,
+                                    op_meta,
+                                    None,
+                                    this_group.transactional
+                                )) {
+                                    events.push(port_and_event);
+                                };
+
+                                if first {
+                                    // first window
+                                    //          swap(scratch, aggrs)
+                                    std::mem::swap(
+                                        &mut scratch_last_window.aggregates,
+                                        &mut this_group.aggrs,
+                                    ); // store current state before init
+                                    scratch_last_window.transactional = this_group.transactional;
+                                    //          aggrs.init() and reset transactional state
+                                    this_group.reset();
                                 } else {
-                                    break;
+                                    // not first window
+                                    // store old aggrs state from last window into scratch2
+
+                                    std::mem::swap(scratch2, scratch_last_window);
+                                    //          swap(aggrs, scratch)
+                                    // store aggrs state before init() into scratch1 for next window
+                                    std::mem::swap(
+                                        &mut scratch_last_window.aggregates,
+                                        &mut this_group.aggrs,
+                                    );
+                                    scratch_last_window.transactional = this_group.transactional;
+                                    //          aggrs.init() and reset transactional state
+                                    // scratch2 <=> scratch_last_window
+                                    // scratch_last_window <=> this
+                                    //
+                                    // scratchg 2 == last window
+                                    // scratch_last_window == this
+                                    // this == empty
+                                    this_group.reset();
+
+                                    //          merge state from last window into this one
+                                    for (this, prev) in
+                                        this_group.aggrs.iter_mut().zip(scratch2.aggregates.iter())
+                                    {
+                                        this.invocable.merge(&prev.invocable).map_err(|e| {
+                                            let r: Option<&Registry> = None;
+                                            e.into_err(prev, prev, r, &node_meta)
+                                        })?;
+                                    }
+                                    this_group.track_transactional(scratch2.transactional);
+                                    this_group.id.track(&event_id_scratch2);
                                 }
                             }
 
-                            // accumulate if we do not emit anything or the first emitting window includes the current event into its state (window_event.include == true)
-                            let pending_accumulate = if let Some(WindowEvent {
-                                emit: true,
-                                include: false,
-                                ..
-                            }) = emit_window_events.first()
-                            {
-                                // postpone
-                                true
-                            } else if let Some(first_window) = windows.first_mut() {
-                                consts.window = Value::from(first_window.name().to_string());
+                            first = false;
+                        }
 
+                        // execute the pending accumulate (if window_event.include == false of first window)
+                        if pending_accumulate {
+                            if let Some(first_window) = windows.first_mut() {
+                                consts.window = Value::from(first_window.name().to_string());
                                 // get current window group
                                 let this_group = stry!(get_or_create_group(
                                     first_window,
@@ -549,245 +715,48 @@ impl Operator for TrickleSelect {
                                 ));
 
                                 // accumulate the current event
-                                let env = Env {
-                                    context: &ctx,
-                                    consts: consts.run(),
-                                    aggrs: &NO_AGGRS,
-                                    meta: &node_meta,
-                                    recursion_limit: tremor_script::recursion_limit(),
-                                };
                                 stry!(accumulate(
+                                    &ctx,
+                                    consts.run(),
                                     opts,
                                     &node_meta,
-                                    &env,
                                     &local_stack,
                                     state,
                                     this_group,
                                     data,
                                     id,
                                     transactional
-                                        ));
-                                false
-                            } else {
-                                // should not happen
-                                false
-                            };
+                                ));
+                            }
+                        }
 
-                            // merge previous aggr state, push window event out, re-initialize for new window
-                            let mut first = true;
-                            let emit_window_iter =
-                                emit_window_events.iter().zip(windows.iter_mut());
-                            for (window_event, window) in emit_window_iter {
-                                consts.window = Value::from(window.name().to_string());
+                        // merge state from last emit window into next non-emit window if there is any
+                        if !emit_window_events.is_empty() {
+                            if let Some(non_emit_window) = windows.get_mut(emit_window_events.len())
+                            {
+                                consts.window = Value::from(non_emit_window.name().to_string());
 
                                 // get current window group
                                 let this_group = stry!(get_or_create_group(
-                                    window,
+                                    non_emit_window,
                                     event_id_gen,
                                     aggregates,
                                     &group_str,
                                     &group_value,
                                 ));
-
-                                if window_event.include {
-                                    // add this event to the aggr state **BEFORE** emit and propagate to next windows
-                                    // merge with scratch
-                                    if !first {
-                                        for (this, prev) in this_group
-                                            .aggrs
-                                            .iter_mut()
-                                            .zip(scratch1.aggregates.iter())
-                                        {
-                                            this.invocable.merge(&prev.invocable).map_err(|e| {
-                                                let r: Option<&Registry> = None;
-                                                e.into_err(prev, prev, r, &node_meta)
-                                            })?;
-                                        }
-                                        this_group.track_transactional(scratch1.transactional);
-                                        this_group.id.track(&event_id_scratch1);
-                                    }
-                                    // push event
-                                    let env = Env {
-                                        context: &ctx,
-                                        consts: consts.run(),
-                                        aggrs: &this_group.aggrs,
-                                        meta: &node_meta,
-                                        recursion_limit: tremor_script::recursion_limit(),
-                                    };
-
-                                    let mut outgoing_event_id = event_id_gen.next_id();
-                                    event_id_scratch1 = this_group.id.clone(); // store the id for to be tracked by the next window
-                                    mem::swap(&mut outgoing_event_id, &mut this_group.id);
-                                    if let Some(port_and_event) = stry!(execute_select_and_having(
-                                        &select,
-                                        &node_meta,
-                                        opts,
-                                        &local_stack,
-                                        state,
-                                        &env,
-                                        outgoing_event_id,
-                                        &data,
-                                        ingest_ns,
-                                        op_meta,
-                                                    None,
-                                        this_group.transactional
-                                    )) {
-                                        events.push(port_and_event);
-                                    };
-
-                                    // swap(aggrs, scratch) - store state before init to scratch as we need it for the next window
-                                    std::mem::swap(&mut scratch1.aggregates, &mut this_group.aggrs);
-                                    // store transactional state for mixing into the next window
-                                    scratch1.transactional = this_group.transactional;
-
-                                    // aggrs.init() and reset transactional state
-                                    this_group.reset();
-                                } else {
-                                    // add this event to the aggr state **AFTER** emit and propagate to next windows
-
-                                    // push event
-                                    let env = Env {
-                                        context: &ctx,
-                                        consts: consts.run(),
-                                        aggrs: &this_group.aggrs,
-                                        meta: &node_meta,
-                                        recursion_limit: tremor_script::recursion_limit(),
-                                    };
-
-                                    let mut outgoing_event_id = event_id_gen.next_id();
-                                    if !first {
-                                        // store old event_ids state from last window into scratch2
-                                        std::mem::swap(
-                                            &mut event_id_scratch1,
-                                            &mut event_id_scratch2,
-                                        );
-                                    }
-                                    event_id_scratch1 = this_group.id.clone(); // store the id for to be tracked by the next window
-                                    mem::swap(&mut outgoing_event_id, &mut this_group.id);
-                                    if let Some(port_and_event) = stry!(execute_select_and_having(
-                                        &select,
-                                        &node_meta,
-                                        opts,
-                                        &local_stack,
-                                        state,
-                                        &env,
-                                        outgoing_event_id,
-                                        &data,
-                                        ingest_ns,
-                                        op_meta,
-                                                    None,
-                                        this_group.transactional
-                                    )) {
-                                        events.push(port_and_event);
-                                    };
-
-                                    if first {
-                                        // first window
-                                        //          swap(scratch, aggrs)
-                                        std::mem::swap(
-                                            &mut scratch1.aggregates,
-                                            &mut this_group.aggrs,
-                                        ); // store current state before init
-                                        scratch1.transactional = this_group.transactional;
-                                        //          aggrs.init() and reset transactional state
-                                        this_group.reset();
-                                    } else {
-                                        // not first window
-                                        // store old aggrs state from last window into scratch2
-                                        std::mem::swap(scratch1, scratch2);
-                                        //          swap(aggrs, scratch) // store aggrs state before init() into scratch1 for next window
-                                        std::mem::swap(
-                                            &mut this_group.aggrs,
-                                            &mut scratch1.aggregates,
-                                        );
-                                        scratch1.transactional = this_group.transactional;
-                                        //          aggrs.init() and reset transactional state
-                                        this_group.reset();
-
-                                        //          merge state from last window into this one
-                                        for (this, prev) in this_group
-                                            .aggrs
-                                            .iter_mut()
-                                            .zip(scratch2.aggregates.iter())
-                                        {
-                                            this.invocable.merge(&prev.invocable).map_err(|e| {
-                                                let r: Option<&Registry> = None;
-                                                e.into_err(prev, prev, r, &node_meta)
-                                            })?;
-                                        }
-                                        this_group.track_transactional(scratch2.transactional);
-                                        this_group.id.track(&event_id_scratch2);
-                                    }
-                                }
-
-                                first = false;
-                            }
-
-                            // execute the pending accumulate (if window_event.include == false of first window)
-                            if pending_accumulate {
-                                if let Some(first_window) = windows.first_mut() {
-                                    consts.window = Value::from(first_window.name().to_string());
-                                    // get current window group
-                                    let this_group = stry!(get_or_create_group(
-                                        first_window,
-                                        event_id_gen,
-                                        aggregates,
-                                        &group_str,
-                                        &group_value,
-                                    ));
-
-                                    // accumulate the current event
-                                    let env = Env {
-                                        context: &ctx,
-                                        consts: consts.run(),
-                                        aggrs: &NO_AGGRS,
-                                        meta: &node_meta,
-                                        recursion_limit: tremor_script::recursion_limit(),
-                                    };
-                                    stry!(accumulate(
-                                        opts,
-                                        &node_meta,
-                                        &env,
-                                        &local_stack,
-                                        state,
-                                        this_group,
-                                        data,
-                                        id,
-                                        transactional
-                                                ));
-                                }
-                            }
-
-                            // merge state from last emit window into next non-emit window if there is any
-                            if !emit_window_events.is_empty() {
-                                if let Some(non_emit_window) =
-                                   windows.get_mut(emit_window_events.len())
+                                for (this, prev) in this_group
+                                    .aggrs
+                                    .iter_mut()
+                                    .zip(scratch_last_window.aggregates.iter())
                                 {
-                                    consts.window = Value::from(non_emit_window.name().to_string());
-
-                                    // get current window group
-                                    let this_group = stry!(get_or_create_group(
-                                        non_emit_window,
-                                        event_id_gen,
-                                        aggregates,
-                                        &group_str,
-                                        &group_value,
-                                    ));
-                                    for (this, prev) in
-                                        this_group.aggrs.iter_mut().zip(scratch1.aggregates.iter())
-                                    {
-                                        this.invocable.merge(&prev.invocable).map_err(|e| {
-                                            let r: Option<&Registry> = None;
-                                            e.into_err(prev, prev, r, &node_meta)
-                                        })?;
-                                    }
-                                    this_group.track_transactional(scratch1.transactional);
-                                    this_group.id.track(&event_id_scratch1);
+                                    this.invocable.merge(&prev.invocable).map_err(|e| {
+                                        let r: Option<&Registry> = None;
+                                        e.into_err(prev, prev, r, &node_meta)
+                                    })?;
                                 }
+                                this_group.track_transactional(scratch_last_window.transactional);
+                                this_group.id.track(&event_id_scratch_last_window);
                             }
-                        } else {
-                            // if this happens, we fucked up internally.
-                            return Err("Internal Error. Cannot execute select with multiple windows, no scratches available. This is bad!".into());
                         }
                     }
                 }
@@ -817,15 +786,17 @@ impl Operator for TrickleSelect {
             ..
         } = self;
 
-        if signal.kind == Some(SignalKind::Tick) && !windows.is_empty() {
-            let ingest_ns = signal.ingest_ns;
-            // Handle eviction
-            // retire the group data that didnt receive an event in `eviction_ns()` nanoseconds
-            // if no event came after `2 * eviction_ns()` this group is finally cleared out
-            windows.iter_mut().for_each(|w| w.maybe_evict(ingest_ns));
+        if signal.kind != Some(SignalKind::Tick) || windows.is_empty() {
+            return Ok(EventAndInsights::default());
+        }
+        let ingest_ns = signal.ingest_ns;
+        // Handle eviction
+        // retire the group data that didnt receive an event in `eviction_ns()` nanoseconds
+        // if no event came after `2 * eviction_ns()` this group is finally cleared out
+        windows.iter_mut().for_each(|w| w.maybe_evict(ingest_ns));
 
-            let opts = Self::opts();
-            select.rent_mut(|stmt| {
+        let opts = Self::opts();
+        select.rent_mut(|stmt| {
             let SelectStmt {
                 stmt,
                 aggregates,
@@ -833,7 +804,7 @@ impl Operator for TrickleSelect {
                 consts,
                 locals,
                 node_meta,
-            }  = stmt;
+            } = stmt;
             let mut res = EventAndInsights::default();
 
             let vm: ValueAndMeta = (Value::null(), Value::object()).into();
@@ -844,1361 +815,267 @@ impl Operator for TrickleSelect {
             consts.group = Value::const_null();
             consts.args = Value::const_null();
 
-            match windows.len() {
-                0 => {} // we shouldnt get here
-                1 => {
-                    let ctx = EventContext::new(signal.ingest_ns, None);
-                    for window in windows.iter_mut() {
-                        consts.window = Value::from(window.name().to_string());
-                        // iterate all groups, including the ones from last_dims
-                        for (group_str, group_data) in window
-                            .dims
-                            .iter_mut()
-                            .chain(window.last_dims.iter_mut())
-                        {
-                            consts.group = group_data.group.clone_static();
-                            consts.group.push(group_str.clone())?;
-
-                            let window_event = stry!(group_data.window.on_tick(ingest_ns));
-                            if window_event.emit {
-                                // evaluate the event and push
-                                let env = Env {
-                                    context: &ctx,
-                                    consts: consts.run(),
-                                    aggrs: &group_data.aggrs,
-                                    meta: &node_meta,
-                                    recursion_limit: tremor_script::recursion_limit(),
-                                };
-                                let mut outgoing_event_id = event_id_gen.next_id();
-                                mem::swap(&mut group_data.id, &mut outgoing_event_id);
-                                let executed = stry!(execute_select_and_having(
-                                    &stmt,
-                                    node_meta,
-                                    opts,
-                                    &local_stack,
-                                    state,
-                                    &env,
-                                    outgoing_event_id,
-                                    &vm,
-                                    ingest_ns,
-                                    &op_meta,
-                                    None,
-                                    group_data.transactional
-                                ));
-                                if let Some(port_and_event) = executed {
-                                    res.events.push(port_and_event);
-                                }
-
-                                // init aggregates and reset transactional status
-                                group_data.reset();
-                            }
-                        }
-                    }
-                }
-                _ => {
-                    // complicated logic
-                    // get a set of groups
-                    //   for each window
-                    //      get the corresponding group data
-                    //      on_event ... do the magic dance with scratches etc
-                    // we need scratches for multiple window handling
-
-                    let ctx = EventContext::new(ingest_ns, None);
-
-                    // we get away with just checking the first window as there can be no group in a further window
-                    // that is not also in the first.
-                    let distinct_groups: Vec<(&String, &Value)> = windows
-                        .first()
-                        .iter()
-                        .flat_map(|window| {
-                            (*window)
-                                .dims
-                                .iter()
-                                .chain(window.last_dims.iter())
-                        })
-                        .map(|(group_str, group_data)| {
-                            // trick the borrow checker, so we can reference self.windows mutably below
-                            // this is safe, as we only ever work on group_data.window and group_data.aggregates
-                            // we might move a group_data struct from one map to another, but there is no deletion
-                            // we never touch neither the group value nor the group_str during the fiddling below
-                            // ALLOW: https://github.com/tremor-rs/tremor-runtime/issues/1027
-                            let group_str: &'static String = unsafe { std::mem::transmute(group_str) };
-                            let group_value: &'static Value =
-                            // ALLOW: https://github.com/tremor-rs/tremor-runtime/issues/1027
-                                unsafe { std::mem::transmute(&group_data.group) }; // lose the reference to window
-                            (group_str, group_value)
-                        })
-                        .collect();
-
-                    for (group_str, group_value) in distinct_groups {
-                        consts.group = group_value.clone_static();
+            let ctx = EventContext::new(ingest_ns, None);
+            if windows.len() == 1 {
+                for window in windows.iter_mut() {
+                    consts.window = Value::from(window.name().to_string());
+                    // iterate all groups, including the ones from last_dims
+                    for (group_str, group_data) in
+                        window.dims.iter_mut().chain(window.last_dims.iter_mut())
+                    {
+                        consts.group = group_data.group.clone_static();
                         consts.group.push(group_str.clone())?;
 
-                        if let Some((scratch1, scratch2)) = aggregate_scratches.as_mut() {
-                            // scratches for passing on the event id between windows to track all events contributing to an outgoing windowed event
-                            let mut event_id_scratch1 = EventId::default();
-                            let mut event_id_scratch2 = EventId::default();
-
-                            // gather window events
-                            let mut emit_window_events = Vec::with_capacity(windows.len());
-                            let step1_iter = windows.iter_mut();
-                            for window in step1_iter {
-                                consts.window = Value::from(window.name().to_string());
-                                // get current window group
-                                let this_group = stry!(get_or_create_group(
-                                    window,
-                                    event_id_gen,
-                                    aggregates,
-                                    group_str,
-                                    group_value,
-                                ));
-                                let window_event = stry!(this_group.window.on_tick(ingest_ns));
-                                if window_event.emit {
-                                    emit_window_events.push(window_event);
-                                } else {
-                                    break;
-                                }
+                        let window_event = stry!(group_data.window.on_tick(ingest_ns));
+                        if window_event.emit {
+                            // evaluate the event and push
+                            let env = Env {
+                                context: &ctx,
+                                consts: consts.run(),
+                                aggrs: &group_data.aggrs,
+                                meta: &node_meta,
+                                recursion_limit: tremor_script::recursion_limit(),
+                            };
+                            let mut outgoing_event_id = event_id_gen.next_id();
+                            mem::swap(&mut group_data.id, &mut outgoing_event_id);
+                            let executed = stry!(execute_select_and_having(
+                                &stmt,
+                                node_meta,
+                                opts,
+                                &local_stack,
+                                state,
+                                &env,
+                                outgoing_event_id,
+                                &vm,
+                                ingest_ns,
+                                &op_meta,
+                                None,
+                                group_data.transactional
+                            ));
+                            if let Some(port_and_event) = executed {
+                                res.events.push(port_and_event);
                             }
 
-                            // merge previous aggr state, push window event out, re-initialize for new window
-                            let mut first = true;
-                            let emit_window_iter =
-                                emit_window_events.iter().zip(windows.iter_mut());
-                            for (window_event, window) in emit_window_iter {
-                                consts.window = Value::from(window.name().to_string());
-
-                                // get current window group
-                                let this_group = stry!(get_or_create_group(
-                                    window,
-                                    event_id_gen,
-                                    aggregates,
-                                    group_str,
-                                    group_value,
-                                ));
-
-                                if window_event.include {
-                                    // add this event to the aggr state **BEFORE** emit and propagate to next windows
-                                    // merge with scratch
-                                    if !first {
-                                        for (this, prev) in this_group
-                                            .aggrs
-                                            .iter_mut()
-                                            .zip(scratch1.aggregates.iter())
-                                        {
-                                            this.invocable.merge(&prev.invocable).map_err(|e| {
-                                                let r: Option<&Registry> = None;
-                                                e.into_err(prev, prev, r, &node_meta)
-                                            })?;
-                                        }
-                                        this_group.track_transactional(scratch1.transactional);
-                                        this_group.id.track(&event_id_scratch1);
-                                    }
-                                    // push event
-                                    let env = Env {
-                                        context: &ctx,
-                                        consts: consts.run(),
-                                        aggrs: &this_group.aggrs,
-                                        meta: &node_meta,
-                                        recursion_limit: tremor_script::recursion_limit(),
-                                    };
-
-                                    let mut outgoing_event_id = event_id_gen.next_id();
-                                    event_id_scratch1 = this_group.id.clone(); // store event_id in a scratch for tracking it in the next window in the tilt-frame
-                                    mem::swap(&mut this_group.id, &mut outgoing_event_id);
-                                    if let Some(port_and_event) = stry!(execute_select_and_having(
-                                        &stmt,
-                                        &node_meta,
-                                        opts,
-                                        &local_stack,
-                                        state,
-                                        &env,
-                                        outgoing_event_id,
-                                        &vm,
-                                        ingest_ns,
-                                        &op_meta,
-                                            None,
-                                        this_group.transactional
-                                    )) {
-                                        res.events.push(port_and_event);
-                                    };
-
-                                    // swap(aggrs, scratch) - store state before init to scratch as we need it for the next window
-                                    std::mem::swap(&mut scratch1.aggregates, &mut this_group.aggrs);
-                                    scratch1.transactional = this_group.transactional;
-                                    // aggrs.init() and reset transactional state
-                                    this_group.reset();
-                                } else {
-                                    // add this event to the aggr state **AFTER** emit and propagate to next windows
-
-                                    // push event
-                                    let env = Env {
-                                        context: &ctx,
-                                        consts: consts.run(),
-                                        aggrs: &this_group.aggrs,
-                                        meta: &node_meta,
-                                        recursion_limit: tremor_script::recursion_limit(),
-                                    };
-
-                                    let mut outgoing_event_id = event_id_gen.next_id();
-                                    if !first {
-                                        // store previous window event_id in scratch2
-                                        std::mem::swap(
-                                            &mut event_id_scratch1,
-                                            &mut event_id_scratch2,
-                                        );
-                                    }
-                                    event_id_scratch1 = this_group.id.clone(); // store event id before swapping it out, for tracking in next window
-                                    mem::swap(&mut this_group.id, &mut outgoing_event_id);
-                                    if let Some(port_and_event) = stry!(execute_select_and_having(
-                                        &stmt,
-                                        &node_meta,
-                                        opts,
-                                        &local_stack,
-                                        state,
-                                        &env,
-                                        outgoing_event_id,
-                                        &vm,
-                                        ingest_ns,
-                                        &op_meta,
-                                            None,
-                                        this_group.transactional
-                                    )) {
-                                        res.events.push(port_and_event);
-                                    };
-
-                                    if first {
-                                        // first window
-                                        //          swap(scratch, aggrs)
-                                        std::mem::swap(
-                                            &mut scratch1.aggregates,
-                                            &mut this_group.aggrs,
-                                        ); // store current state before init
-                                        scratch1.transactional = this_group.transactional;
-                                        //          aggrs.init() and reset transactional state
-                                        this_group.reset();
-                                    } else {
-                                        // not first window
-                                        // store old aggrs state from last window into scratch2
-                                        std::mem::swap(scratch1, scratch2);
-                                        //          swap(aggrs, scratch) // store aggrs state before init() into scratch1 for next window
-                                        std::mem::swap(
-                                            &mut this_group.aggrs,
-                                            &mut scratch1.aggregates,
-                                        );
-                                        scratch1.transactional = this_group.transactional;
-
-                                        //          aggrs.init() and reset transactional state
-                                        this_group.reset();
-                                        //          merge state from last window into this one
-                                        for (this, prev) in this_group
-                                            .aggrs
-                                            .iter_mut()
-                                            .zip(scratch2.aggregates.iter())
-                                        {
-                                            this.invocable.merge(&prev.invocable).map_err(|e| {
-                                                let r: Option<&Registry> = None;
-                                                e.into_err(prev, prev, r, &node_meta)
-                                            })?;
-                                        }
-                                        this_group.track_transactional(scratch2.transactional);
-                                        this_group.id.track(&event_id_scratch2);
-                                    }
-                                }
-
-                                first = false;
-                            }
-
-                            // merge state from last emit window into next non-emit window if there is any
-                            if !emit_window_events.is_empty() {
-                                if let Some(non_emit_window) =
-                                    windows.get_mut(emit_window_events.len())
-                                {
-                                    consts.window = Value::from(non_emit_window.name().to_string());
-
-                                    // get current window group
-                                    let this_group = stry!(get_or_create_group(
-                                        non_emit_window,
-                                        event_id_gen,
-                                        aggregates,
-                                        group_str,
-                                        group_value,
-                                    ));
-                                    for (this, prev) in
-                                        this_group.aggrs.iter_mut().zip(scratch1.aggregates.iter())
-                                    {
-                                        this.invocable.merge(&prev.invocable).map_err(|e| {
-                                            let r: Option<&Registry> = None;
-                                            e.into_err(prev, prev, r, &node_meta)
-                                        })?;
-                                    }
-                                    this_group.track_transactional(scratch1.transactional);
-                                    this_group.id.track(&event_id_scratch1);
-                                }
-                            }
-                        } else {
-                            // if this happens, we fucked up internally.
-                            return Err("Internal Error. Cannot execute select with multiple windows, no scratches available. This is bad!".into());
+                            // init aggregates and reset transactional status
+                            group_data.reset();
                         }
                     }
                 }
-            }
+            } else {
+                // complicated logic
+                // get a set of groups
+                //   for each window
+                //      get the corresponding group data
+                //      on_event ... do the magic dance with scratches etc
+                // we need scratches for multiple window handling
+
+                // we get away with just checking the first window as there can be no group in a further window
+                // that is not also in the first.
+                let distinct_groups: Vec<(String, Value)> = windows
+                    .first()
+                    .iter()
+                    .flat_map(|window| (*window).dims.iter().chain(window.last_dims.iter()))
+                    .map(|(group_str, group_data)| {
+                        (group_str.clone(), group_data.group.clone_static())
+                    })
+                    .collect();
+
+                for (group_str, group_value) in distinct_groups {
+                    consts.group = group_value.clone();
+                    consts.group.push(group_str.clone())?;
+
+                    let (scratch1, scratch2) = aggregate_scratches;
+                    // scratches for passing on the event id between windows to track all events contributing to an outgoing windowed event
+                    let mut event_id_scratch1 = EventId::default();
+                    let mut event_id_scratch2 = EventId::default();
+
+                    // gather window events
+                    let mut emit_window_events = Vec::with_capacity(windows.len());
+                    let step1_iter = windows.iter_mut();
+                    for window in step1_iter {
+                        consts.window = Value::from(window.name().to_string());
+                        // get current window group
+                        let this_group = stry!(get_or_create_group(
+                            window,
+                            event_id_gen,
+                            aggregates,
+                            &group_str,
+                            &group_value,
+                        ));
+                        let window_event = stry!(this_group.window.on_tick(ingest_ns));
+                        if window_event.emit {
+                            emit_window_events.push(window_event);
+                        } else {
+                            break;
+                        }
+                    }
+
+                    // merge previous aggr state, push window event out, re-initialize for new window
+                    let mut first = true;
+                    let emit_window_iter = emit_window_events.iter().zip(windows.iter_mut());
+                    for (window_event, window) in emit_window_iter {
+                        consts.window = Value::from(window.name().to_string());
+
+                        // get current window group
+                        let this_group = stry!(get_or_create_group(
+                            window,
+                            event_id_gen,
+                            aggregates,
+                            &group_str,
+                            &group_value,
+                        ));
+
+                        if window_event.include {
+                            // add this event to the aggr state **BEFORE** emit and propagate to next windows
+                            // merge with scratch
+                            if !first {
+                                for (this, prev) in
+                                    this_group.aggrs.iter_mut().zip(scratch1.aggregates.iter())
+                                {
+                                    this.invocable.merge(&prev.invocable).map_err(|e| {
+                                        let r: Option<&Registry> = None;
+                                        e.into_err(prev, prev, r, &node_meta)
+                                    })?;
+                                }
+                                this_group.track_transactional(scratch1.transactional);
+                                this_group.id.track(&event_id_scratch1);
+                            }
+                            // push event
+                            let env = Env {
+                                context: &ctx,
+                                consts: consts.run(),
+                                aggrs: &this_group.aggrs,
+                                meta: &node_meta,
+                                recursion_limit: tremor_script::recursion_limit(),
+                            };
+
+                            let mut outgoing_event_id = event_id_gen.next_id();
+                            event_id_scratch1 = this_group.id.clone(); // store event_id in a scratch for tracking it in the next window in the tilt-frame
+                            mem::swap(&mut this_group.id, &mut outgoing_event_id);
+                            if let Some(port_and_event) = stry!(execute_select_and_having(
+                                &stmt,
+                                &node_meta,
+                                opts,
+                                &local_stack,
+                                state,
+                                &env,
+                                outgoing_event_id,
+                                &vm,
+                                ingest_ns,
+                                &op_meta,
+                                None,
+                                this_group.transactional
+                            )) {
+                                res.events.push(port_and_event);
+                            };
+
+                            // swap(aggrs, scratch) - store state before init to scratch as we need it for the next window
+                            std::mem::swap(&mut scratch1.aggregates, &mut this_group.aggrs);
+                            scratch1.transactional = this_group.transactional;
+                            // aggrs.init() and reset transactional state
+                            this_group.reset();
+                        } else {
+                            // add this event to the aggr state **AFTER** emit and propagate to next windows
+
+                            // push event
+                            let env = Env {
+                                context: &ctx,
+                                consts: consts.run(),
+                                aggrs: &this_group.aggrs,
+                                meta: &node_meta,
+                                recursion_limit: tremor_script::recursion_limit(),
+                            };
+
+                            let mut outgoing_event_id = event_id_gen.next_id();
+                            if !first {
+                                // store previous window event_id in scratch2
+                                std::mem::swap(&mut event_id_scratch1, &mut event_id_scratch2);
+                            }
+                            event_id_scratch1 = this_group.id.clone(); // store event id before swapping it out, for tracking in next window
+                            mem::swap(&mut this_group.id, &mut outgoing_event_id);
+                            if let Some(port_and_event) = stry!(execute_select_and_having(
+                                &stmt,
+                                &node_meta,
+                                opts,
+                                &local_stack,
+                                state,
+                                &env,
+                                outgoing_event_id,
+                                &vm,
+                                ingest_ns,
+                                &op_meta,
+                                None,
+                                this_group.transactional
+                            )) {
+                                res.events.push(port_and_event);
+                            };
+
+                            if first {
+                                // first window
+                                //          swap(scratch, aggrs)
+                                std::mem::swap(&mut scratch1.aggregates, &mut this_group.aggrs); // store current state before init
+                                scratch1.transactional = this_group.transactional;
+                                //          aggrs.init() and reset transactional state
+                                this_group.reset();
+                            } else {
+                                // not first window
+                                // store old aggrs state from last window into scratch2
+                                std::mem::swap(scratch1, scratch2);
+                                //          swap(aggrs, scratch) // store aggrs state before init() into scratch1 for next window
+                                std::mem::swap(&mut this_group.aggrs, &mut scratch1.aggregates);
+                                scratch1.transactional = this_group.transactional;
+
+                                //          aggrs.init() and reset transactional state
+                                this_group.reset();
+                                //          merge state from last window into this one
+                                for (this, prev) in
+                                    this_group.aggrs.iter_mut().zip(scratch2.aggregates.iter())
+                                {
+                                    this.invocable.merge(&prev.invocable).map_err(|e| {
+                                        let r: Option<&Registry> = None;
+                                        e.into_err(prev, prev, r, &node_meta)
+                                    })?;
+                                }
+                                this_group.track_transactional(scratch2.transactional);
+                                this_group.id.track(&event_id_scratch2);
+                            }
+                        }
+
+                        first = false;
+                    }
+
+                    // merge state from last emit window into next non-emit window if there is any
+                    if !emit_window_events.is_empty() {
+                        if let Some(non_emit_window) = windows.get_mut(emit_window_events.len()) {
+                            consts.window = Value::from(non_emit_window.name().to_string());
+
+                            // get current window group
+                            let this_group = stry!(get_or_create_group(
+                                non_emit_window,
+                                event_id_gen,
+                                aggregates,
+                                &group_str,
+                                &group_value,
+                            ));
+                            for (this, prev) in
+                                this_group.aggrs.iter_mut().zip(scratch1.aggregates.iter())
+                            {
+                                this.invocable.merge(&prev.invocable).map_err(|e| {
+                                    let r: Option<&Registry> = None;
+                                    e.into_err(prev, prev, r, &node_meta)
+                                })?;
+                            }
+                            this_group.track_transactional(scratch1.transactional);
+                            this_group.id.track(&event_id_scratch1);
+                        }
+                    }
+                }
+            };
             Ok(res)
         })
-        } else {
-            Ok(EventAndInsights::default())
-        }
     }
 
     fn handles_signal(&self) -> bool {
         true
-    }
-}
-
-#[cfg(test)]
-mod test {
-
-    #![allow(clippy::float_cmp)]
-    use crate::query::window_decl_to_impl;
-
-    use super::*;
-
-    use tremor_script::ast::{Stmt, WindowDecl};
-    use tremor_script::{
-        ast::{self, Ident, ImutExpr, Literal},
-        path::ModulePath,
-    };
-    use tremor_script::{
-        ast::{AggregateScratch, Consts},
-        Value,
-    };
-    use tremor_value::literal;
-
-    fn test_target<'test>() -> ast::ImutExpr<'test> {
-        let target: ast::ImutExpr<'test> = ImutExpr::from(ast::Literal {
-            mid: 0,
-            value: Value::from(42),
-        });
-        target
-    }
-
-    fn test_stmt(target: ast::ImutExpr) -> ast::Select {
-        tremor_script::ast::Select {
-            mid: 0,
-            from: (Ident::from("in"), Ident::from("out")),
-            into: (Ident::from("out"), Ident::from("in")),
-            target,
-            maybe_where: Some(ImutExpr::from(ast::Literal {
-                mid: 0,
-                value: Value::from(true),
-            })),
-            windows: vec![],
-            maybe_group_by: None,
-            maybe_having: None,
-        }
-    }
-
-    fn test_query(stmt: ast::Stmt) -> ast::Query {
-        ast::Query {
-            stmts: vec![stmt.clone()],
-            node_meta: ast::NodeMetas::new(Vec::new()),
-            windows: HashMap::new(),
-            scripts: HashMap::new(),
-            operators: HashMap::new(),
-            config: HashMap::new(),
-        }
-    }
-
-    fn ingest_ns(s: u64) -> u64 {
-        s * 1_000_000_000
-    }
-    fn test_event(s: u64) -> Event {
-        Event {
-            id: (0, 0, s).into(),
-            ingest_ns: ingest_ns(s),
-            data: literal!({
-               "h2g2" : 42,
-            })
-            .into(),
-            ..Event::default()
-        }
-    }
-
-    fn test_event_tx(s: u64, transactional: bool, group: u64) -> Event {
-        Event {
-            id: (0, 0, s).into(),
-            ingest_ns: s + 100,
-            data: literal!({ "group": group }).into(),
-            transactional,
-            ..Event::default()
-        }
-    }
-
-    fn test_select(uid: u64, stmt: srs::Stmt) -> Result<TrickleSelect> {
-        let groups = window::Groups::new();
-        let windows = vec![
-            (
-                "w15s".into(),
-                window::TumblingOnTime {
-                    max_groups: window::Impl::DEFAULT_MAX_GROUPS,
-                    interval: 15_000_000_000,
-                    ..Default::default()
-                }
-                .into(),
-            ),
-            (
-                "w30s".into(),
-                window::TumblingOnTime {
-                    emit_empty_windows: false,
-                    ttl: 30_000_000_000,
-                    max_groups: window::Impl::DEFAULT_MAX_GROUPS,
-                    interval: 30_000_000_000,
-                    ..Default::default()
-                }
-                .into(),
-            ),
-        ];
-        let id = "select".to_string();
-        TrickleSelect::with_stmt(uid, id, &groups, windows, &stmt)
-    }
-
-    fn try_enqueue(
-        op: &mut TrickleSelect,
-        event: Event,
-    ) -> Result<Option<(Cow<'static, str>, Event)>> {
-        let mut state = Value::null();
-        let mut action = op.on_event(0, "in", &mut state, event)?;
-        let first = action.events.pop();
-        if action.events.is_empty() {
-            Ok(first)
-        } else {
-            Ok(None)
-        }
-    }
-
-    fn try_enqueue_two(
-        op: &mut TrickleSelect,
-        event: Event,
-    ) -> Result<Option<[(Cow<'static, str>, Event); 2]>> {
-        let mut state = Value::null();
-        let mut action = op.on_event(0, "in", &mut state, event)?;
-        let r = action
-            .events
-            .pop()
-            .and_then(|second| Some([action.events.pop()?, second]));
-        if action.events.is_empty() {
-            Ok(r)
-        } else {
-            Ok(None)
-        }
-    }
-
-    fn parse_query(
-        file_name: String,
-        query: &str,
-    ) -> Result<crate::op::trickle::select::TrickleSelect> {
-        let reg = tremor_script::registry();
-        let aggr_reg = tremor_script::aggr_registry();
-        let module_path = tremor_script::path::load();
-        let cus = vec![];
-        let query = tremor_script::query::Query::parse(
-            &module_path,
-            &file_name,
-            query,
-            cus,
-            &reg,
-            &aggr_reg,
-        )
-        .map_err(tremor_script::errors::CompilerError::error)?;
-        let stmt = srs::Stmt::try_new_from_query(&query.query, |q| {
-            q.stmts
-                .first()
-                .cloned()
-                .ok_or_else(|| Error::from("Invalid query"))
-        })?;
-        Ok(test_select(1, stmt)?)
-    }
-
-    #[test]
-    fn test_sum() -> Result<()> {
-        let mut op = parse_query(
-            "test.trickle".to_string(),
-            "select aggr::stats::sum(event.h2g2) from in[w15s, w30s] into out;",
-        )?;
-        assert!(try_enqueue(&mut op, test_event(0))?.is_none());
-        assert!(try_enqueue(&mut op, test_event(1))?.is_none());
-        let (out, event) =
-            try_enqueue(&mut op, test_event(15))?.expect("no event emitted after aggregation");
-        assert_eq!("out", out);
-
-        assert_eq!(*event.data.suffix().value(), 84.0);
-        Ok(())
-    }
-
-    #[test]
-    fn test_count() -> Result<()> {
-        let mut op = parse_query(
-            "test.trickle".to_string(),
-            "select aggr::stats::count() from in[w15s, w30s] into out;",
-        )?;
-        assert!(try_enqueue(&mut op, test_event(0))?.is_none());
-        assert!(try_enqueue(&mut op, test_event(1))?.is_none());
-        let (out, event) = try_enqueue(&mut op, test_event(15))?.expect("no event");
-        assert_eq!("out", out);
-
-        assert_eq!(*event.data.suffix().value(), 2);
-        Ok(())
-    }
-
-    fn select_stmt_from_query(query_str: &str) -> Result<TrickleSelect> {
-        let reg = tremor_script::registry();
-        let aggr_reg = tremor_script::aggr_registry();
-        let module_path = tremor_script::path::load();
-        let cus = vec![];
-        let query = tremor_script::query::Query::parse(
-            &module_path,
-            "fake",
-            query_str,
-            cus,
-            &reg,
-            &aggr_reg,
-        )
-        .map_err(tremor_script::errors::CompilerError::error)?;
-        let window_decls: Vec<WindowDecl<'_>> = query
-            .suffix()
-            .stmts
-            .iter()
-            .filter_map(|stmt| match stmt {
-                Stmt::WindowDecl(wd) => Some(wd.as_ref().clone()),
-                _ => None,
-            })
-            .collect();
-        let stmt = srs::Stmt::try_new_from_query(&query.query, |q| {
-            q.stmts
-                .iter()
-                .find(|stmt| matches!(*stmt, Stmt::Select(_)))
-                .cloned()
-                .ok_or_else(|| Error::from("Invalid query, expected only 1 select statement"))
-        })?;
-        let windows: Vec<(String, window::Impl)> = window_decls
-            .iter()
-            .enumerate()
-            .map(|(i, window_decl)| {
-                (
-                    i.to_string(),
-                    window_decl_to_impl(window_decl).unwrap(), // yes, indeed!
-                )
-            })
-            .collect();
-        let groups = window::Groups::new();
-
-        let id = "select".to_string();
-        Ok(TrickleSelect::with_stmt(42, id, &groups, windows, &stmt)?)
-    }
-
-    fn test_tick(ns: u64) -> Event {
-        Event {
-            id: EventId::new(1, 1, ns),
-            kind: Some(SignalKind::Tick),
-            ingest_ns: ns,
-            ..Event::default()
-        }
-    }
-
-    #[test]
-    fn select_single_win_with_script_on_signal() -> Result<()> {
-        let mut select = select_stmt_from_query(
-            r#"
-        define tumbling window window1
-        with
-            interval = 5
-        script
-            event.time
-        end;
-        select aggr::stats::count() from in[window1] group by event.g into out having event > 0;
-        "#,
-        )?;
-        let uid = 42;
-        let mut state = Value::null();
-        let mut tick1 = test_tick(1);
-        let mut eis = select.on_signal(uid, &state, &mut tick1)?;
-        assert!(eis.insights.is_empty());
-        assert_eq!(0, eis.events.len());
-
-        let event = Event {
-            id: (1, 1, 300).into(),
-            ingest_ns: 1_000,
-            data: literal!({
-               "time" : 4,
-               "g": "group"
-            })
-            .into(),
-            ..Event::default()
-        };
-        eis = select.on_event(uid, "IN", &mut state, event)?;
-        assert!(eis.insights.is_empty());
-        assert_eq!(0, eis.events.len());
-
-        // no emit on signal, although window interval would be passed
-        let mut tick2 = test_tick(10);
-        eis = select.on_signal(uid, &state, &mut tick2)?;
-        assert!(eis.insights.is_empty());
-        assert_eq!(0, eis.events.len());
-
-        // only emit on event
-        let event = Event {
-            id: (1, 1, 300).into(),
-            ingest_ns: 3_000,
-            data: literal!({
-               "time" : 11,
-               "g": "group"
-            })
-            .into(),
-            ..Event::default()
-        };
-        eis = select.on_event(uid, "IN", &mut state, event)?;
-        assert!(eis.insights.is_empty());
-        assert_eq!(1, eis.events.len());
-        assert_eq!("1", sorted_serialize(eis.events[0].1.data.parts().0)?);
-        Ok(())
-    }
-
-    #[test]
-    fn select_single_win_on_signal() -> Result<()> {
-        let mut select = select_stmt_from_query(
-            r#"
-        define tumbling window window1
-        with
-            interval = 2
-        end;
-        select aggr::win::collect_flattened(event) from in[window1] group by event.g into out;
-        "#,
-        )?;
-        let uid = 42;
-        let mut state = Value::null();
-        let mut tick1 = test_tick(1);
-        let mut eis = select.on_signal(uid, &state, &mut tick1)?;
-        assert!(eis.insights.is_empty());
-        assert_eq!(0, eis.events.len());
-
-        let event = Event {
-            id: (1, 1, 300).into(),
-            ingest_ns: 2,
-            data: literal!({
-               "g": "group"
-            })
-            .into(),
-            ..Event::default()
-        };
-        eis = select.on_event(uid, "IN", &mut state, event)?;
-        assert!(eis.insights.is_empty());
-        assert_eq!(0, eis.events.len());
-
-        // no emit yet
-        let mut tick2 = test_tick(3);
-        eis = select.on_signal(uid, &state, &mut tick2)?;
-        assert!(eis.insights.is_empty());
-        assert_eq!(0, eis.events.len());
-
-        // now emit
-        let mut tick3 = test_tick(4);
-        eis = select.on_signal(uid, &state, &mut tick3)?;
-        assert!(eis.insights.is_empty());
-        assert_eq!(1, eis.events.len());
-        assert_eq!(
-            r#"[{"g":"group"}]"#,
-            sorted_serialize(eis.events[0].1.data.parts().0)?
-        );
-        assert_eq!(false, eis.events[0].1.transactional);
-        Ok(())
-    }
-
-    #[test]
-    fn select_multiple_wins_on_signal() -> Result<()> {
-        let mut select = select_stmt_from_query(
-            r#"
-        define tumbling window window1
-        with
-            interval = 100
-        end;
-        define tumbling window window2
-        with
-            size = 2
-        end;
-        select aggr::win::collect_flattened(event) from in[window1, window2] group by event.cat into out;
-        "#,
-        )?;
-        let uid = 42;
-        let mut state = Value::null();
-        let mut tick1 = test_tick(1);
-        let mut eis = select.on_signal(uid, &state, &mut tick1)?;
-        assert!(eis.insights.is_empty());
-        assert_eq!(0, eis.events.len());
-
-        let mut tick2 = test_tick(100);
-        eis = select.on_signal(uid, &state, &mut tick2)?;
-        assert!(eis.insights.is_empty());
-        assert_eq!(0, eis.events.len());
-
-        // we have no groups, so no emit yet
-        let mut tick3 = test_tick(201);
-        eis = select.on_signal(uid, &state, &mut tick3)?;
-        assert!(eis.insights.is_empty());
-        assert_eq!(0, eis.events.len());
-
-        // insert first event
-        let event = Event {
-            id: (1, 1, 300).into(),
-            ingest_ns: 300,
-            data: literal!({
-               "cat" : 42,
-            })
-            .into(),
-            transactional: true,
-            ..Event::default()
-        };
-        eis = select.on_event(uid, "in", &mut state, event)?;
-        assert!(eis.insights.is_empty());
-        assert_eq!(0, eis.events.len());
-
-        // finish the window with the next tick
-        let mut tick4 = test_tick(401);
-        eis = select.on_signal(uid, &state, &mut tick4)?;
-        assert!(eis.insights.is_empty());
-        assert_eq!(1, eis.events.len());
-        let (_port, event) = dbg!(eis).events.remove(0);
-        assert_eq!(r#"[{"cat":42}]"#, sorted_serialize(event.data.parts().0)?);
-        assert_eq!(true, event.transactional);
-
-        let mut tick5 = test_tick(499);
-        eis = select.on_signal(uid, &state, &mut tick5)?;
-        assert!(eis.insights.is_empty());
-        assert_eq!(0, eis.events.len());
-
-        Ok(())
-    }
-
-    #[test]
-    fn test_transactional_single_window() -> Result<()> {
-        let mut op = select_stmt_from_query(
-            r#"
-            define tumbling window w2
-            with
-              size = 2
-            end;
-            select aggr::stats::count() from in[w2] into out;
-        "#,
-        )?;
-        let mut state = Value::null();
-        let event1 = test_event_tx(0, false, 0);
-        let id1 = event1.id.clone();
-        let res = op.on_event(0, "in", &mut state, event1)?;
-
-        assert!(res.events.is_empty());
-
-        let event2 = test_event_tx(1, true, 0);
-        let id2 = event2.id.clone();
-        let mut res = op.on_event(0, "in", &mut state, event2)?;
-        assert_eq!(1, res.events.len());
-        let (_, event) = res.events.pop().unwrap();
-        assert_eq!(true, event.transactional);
-        assert_eq!(true, event.id.is_tracking(&id1));
-        assert_eq!(true, event.id.is_tracking(&id2));
-
-        let event3 = test_event_tx(2, false, 0);
-        let id3 = event3.id.clone();
-        let res = op.on_event(0, "in", &mut state, event3)?;
-        assert!(res.events.is_empty());
-
-        let event4 = test_event_tx(3, false, 0);
-        let id4 = event4.id.clone();
-        let mut res = op.on_event(0, "in", &mut state, event4)?;
-        assert_eq!(1, res.events.len());
-        let (_, event) = res.events.pop().unwrap();
-        assert_eq!(false, event.transactional);
-        assert_eq!(true, event.id.is_tracking(&id3));
-        assert_eq!(true, event.id.is_tracking(&id4));
-
-        Ok(())
-    }
-
-    #[test]
-    fn test_transactional_multiple_windows() -> Result<()> {
-        let mut op = select_stmt_from_query(
-            r#"
-            define tumbling window w2_1
-            with
-              size = 2
-            end;
-            define tumbling window w2_2
-            with
-              size = 2
-            end;
-            select aggr::win::collect_flattened(event) from in[w2_1, w2_2] group by set(event["group"]) into out;
-        "#,
-        )?;
-        let mut state = Value::null();
-        let event1 = test_event_tx(0, true, 0);
-        let id1 = event1.id.clone();
-        let res = op.on_event(0, "in", &mut state, event1)?;
-        assert_eq!(0, res.len());
-
-        let event2 = test_event_tx(1, false, 1);
-        let id2 = event2.id.clone();
-        let res = op.on_event(0, "in", &mut state, event2)?;
-        assert_eq!(0, res.len());
-
-        let event3 = test_event_tx(2, false, 0);
-        let id3 = event3.id.clone();
-        let mut res = op.on_event(0, "in", &mut state, event3)?;
-        assert_eq!(1, res.len());
-        let (_, event) = res.events.pop().unwrap();
-        assert_eq!(true, event.transactional);
-        assert_eq!(true, event.id.is_tracking(&id1));
-        assert_eq!(true, event.id.is_tracking(&id3));
-
-        let event4 = test_event_tx(3, false, 1);
-        let id4 = event4.id.clone();
-        let mut res = op.on_event(0, "in", &mut state, event4)?;
-        assert_eq!(1, res.len());
-        let (_, event) = res.events.remove(0);
-        assert_eq!(false, event.transactional);
-        assert_eq!(true, event.id.is_tracking(&id2));
-        assert_eq!(true, event.id.is_tracking(&id4));
-
-        let event5 = test_event_tx(4, false, 0);
-        let id5 = event5.id.clone();
-        let res = op.on_event(0, "in", &mut state, event5)?;
-        assert_eq!(0, res.len());
-
-        let event6 = test_event_tx(5, false, 0);
-        let id6 = event6.id.clone();
-        let mut res = op.on_event(0, "in", &mut state, event6)?;
-        assert_eq!(2, res.len());
-
-        // first event from event5 and event6 - none of the source events are transactional
-        let (_, event_w1) = res.events.remove(0);
-        assert_eq!(false, event_w1.transactional);
-        assert_eq!(true, event_w1.id.is_tracking(&id5));
-        assert_eq!(true, event_w1.id.is_tracking(&id6));
-
-        let (_, event_w2) = res.events.remove(0);
-        assert_eq!(true, event_w2.transactional);
-        assert_eq!(true, event_w2.id.is_tracking(&id1));
-        assert_eq!(true, event_w2.id.is_tracking(&id3));
-        assert_eq!(true, event_w2.id.is_tracking(&id5));
-        assert_eq!(true, event_w2.id.is_tracking(&id6));
-
-        Ok(())
-    }
-
-    #[test]
-    fn count_tilt() -> Result<()> {
-        // Windows are 15s and 30s
-        let mut op = select_stmt_from_query(
-            r#"
-        define tumbling window w15s
-        with
-          interval = 15 * 1000000000
-        end;
-        define tumbling window w30s
-        with
-          interval = 30 * 1000000000
-        end;
-        select aggr::stats::count() from in [w15s, w30s] into out;
-        "#,
-        )?;
-
-        // Insert two events prior to 15
-        assert!(try_enqueue(&mut op, test_event(0))?.is_none());
-        assert!(try_enqueue(&mut op, test_event(1))?.is_none());
-        // Add one event at 15, this flushes the prior two and set this to one.
-        // This is the first time the second frame sees an event so it's timer
-        // starts at 15.
-        let (out, event) = try_enqueue(&mut op, test_event(15))?.expect("no event 1");
-        assert_eq!("out", out);
-        assert_eq!(*event.data.suffix().value(), 2);
-        // Add another event prior to 30
-        assert!(try_enqueue(&mut op, test_event(16))?.is_none());
-        // This emits only the initial event since the rollup
-        // will only be emitted once it gets the first event
-        // of the next window
-        let (out, event) = try_enqueue(&mut op, test_event(30))?.expect("no event 2");
-        assert_eq!("out", out);
-        assert_eq!(*event.data.suffix().value(), 2);
-        // Add another event prior to 45 to the first window
-        assert!(try_enqueue(&mut op, test_event(31))?.is_none());
-
-        // At 45 the 15s window emits the rollup with the previous data
-        let [(out1, event1), (out2, event2)] =
-            try_enqueue_two(&mut op, test_event(45))?.expect("no event 3");
-
-        assert_eq!("out", out1);
-        assert_eq!("out", out2);
-        assert_eq!(*event1.data.suffix().value(), 2);
-        assert_eq!(*event2.data.suffix().value(), 4);
-        assert!(try_enqueue(&mut op, test_event(46))?.is_none());
-
-        // Add 60 only the 15s window emits
-        let (out, event) = try_enqueue(&mut op, test_event(60))?.expect("no event 4");
-        assert_eq!("out", out);
-        assert_eq!(*event.data.suffix().value(), 2);
-        Ok(())
-    }
-
-    #[test]
-    fn select_nowin_nogrp_nowhr_nohav() -> Result<()> {
-        let target = test_target();
-        let stmt_ast = test_stmt(target);
-
-        let stmt_ast = test_select_stmt(stmt_ast);
-        let script = "fake".to_string();
-        let query = srs::Query::try_new::<Error, _>(script, |_| Ok(test_query(stmt_ast.clone())))?;
-
-        let stmt = srs::Stmt::try_new_from_query::<Error, _>(&query, |_| Ok(stmt_ast))?;
-
-        let mut op = test_select(1, stmt)?;
-        assert!(try_enqueue(&mut op, test_event(0))?.is_none());
-        assert!(try_enqueue(&mut op, test_event(1))?.is_none());
-        let (out, event) = try_enqueue(&mut op, test_event(15))?.expect("no event");
-        assert_eq!("out", out);
-
-        assert_eq!(*event.data.suffix().value(), 42);
-        Ok(())
-    }
-
-    #[test]
-    fn select_nowin_nogrp_whrt_nohav() -> Result<()> {
-        let target = test_target();
-        let mut stmt_ast = test_stmt(target);
-
-        stmt_ast.maybe_where = Some(ImutExpr::from(ast::Literal {
-            mid: 0,
-            value: Value::from(true),
-        }));
-
-        let stmt_ast = test_select_stmt(stmt_ast);
-        let script = "fake".to_string();
-        let query = srs::Query::try_new::<Error, _>(script, |_| Ok(test_query(stmt_ast.clone())))?;
-
-        let stmt = srs::Stmt::try_new_from_query::<Error, _>(&query, |_| Ok(stmt_ast))?;
-
-        let mut op = test_select(2, stmt)?;
-
-        assert!(try_enqueue(&mut op, test_event(0))?.is_none());
-
-        let (out, event) = try_enqueue(&mut op, test_event(15))?.expect("no event");
-        assert_eq!("out", out);
-
-        assert_eq!(*event.data.suffix().value(), 42);
-        Ok(())
-    }
-
-    #[test]
-    fn select_nowin_nogrp_whrf_nohav() -> Result<()> {
-        let target = test_target();
-        let mut stmt_ast = test_stmt(target);
-
-        let script = "fake".to_string();
-        stmt_ast.maybe_where = Some(ImutExpr::from(ast::Literal {
-            mid: 0,
-            value: Value::from(false),
-        }));
-        let stmt_ast = test_select_stmt(stmt_ast);
-        let query = srs::Query::try_new::<Error, _>(script, |_| Ok(test_query(stmt_ast.clone())))?;
-
-        let stmt = srs::Stmt::try_new_from_query::<Error, _>(&query, |_| Ok(stmt_ast))?;
-
-        let mut op = test_select(3, stmt)?;
-        let next = try_enqueue(&mut op, test_event(0))?;
-        assert_eq!(None, next);
-        Ok(())
-    }
-
-    #[test]
-    fn select_nowin_nogrp_whrbad_nohav() -> Result<()> {
-        let target = test_target();
-        let mut stmt_ast = test_stmt(target);
-        stmt_ast.maybe_where = Some(ImutExpr::from(ast::Literal {
-            mid: 0,
-            value: Value::from("snot"),
-        }));
-
-        let stmt_ast = test_select_stmt(stmt_ast);
-        let script = "fake".to_string();
-        let query = srs::Query::try_new::<Error, _>(script, |_| Ok(test_query(stmt_ast.clone())))?;
-
-        let stmt = srs::Stmt::try_new_from_query::<Error, _>(&query, |_| Ok(stmt_ast))?;
-
-        let mut op = test_select(4, stmt)?;
-
-        assert!(try_enqueue(&mut op, test_event(0)).is_err());
-
-        Ok(())
-    }
-
-    #[test]
-    fn select_nowin_nogrp_whrt_havt() -> Result<()> {
-        let target = test_target();
-        let mut stmt_ast = test_stmt(target);
-        stmt_ast.maybe_where = Some(ImutExpr::from(ast::Literal {
-            mid: 0,
-            value: Value::from(true),
-        }));
-        stmt_ast.maybe_having = Some(ImutExpr::from(ast::Literal {
-            mid: 0,
-            value: Value::from(true),
-        }));
-
-        let stmt_ast = test_select_stmt(stmt_ast);
-        let script = "fake".to_string();
-        let query = srs::Query::try_new::<Error, _>(script, |_| Ok(test_query(stmt_ast.clone())))?;
-
-        let stmt = srs::Stmt::try_new_from_query::<Error, _>(&query, |_| Ok(stmt_ast))?;
-
-        let mut op = test_select(5, stmt)?;
-
-        let event = test_event(0);
-        assert!(try_enqueue(&mut op, event)?.is_none());
-
-        let event = test_event(15);
-        let (out, event) = try_enqueue(&mut op, event)?.expect("no event");
-        assert_eq!("out", out);
-        assert_eq!(*event.data.suffix().value(), 42);
-        Ok(())
-    }
-
-    #[test]
-    fn select_nowin_nogrp_whrt_havf() -> Result<()> {
-        let target = test_target();
-        let mut stmt_ast = test_stmt(target);
-        stmt_ast.maybe_where = Some(ImutExpr::from(ast::Literal {
-            mid: 0,
-            value: Value::from(true),
-        }));
-        stmt_ast.maybe_having = Some(ImutExpr::from(ast::Literal {
-            mid: 0,
-            value: Value::from(false),
-        }));
-
-        let stmt_ast = test_select_stmt(stmt_ast);
-        let script = "fake".to_string();
-        let query = srs::Query::try_new::<Error, _>(script, |_| Ok(test_query(stmt_ast.clone())))?;
-
-        let stmt = srs::Stmt::try_new_from_query::<Error, _>(&query, |_| Ok(stmt_ast))?;
-
-        let mut op = test_select(6, stmt)?;
-        let event = test_event(0);
-
-        let next = try_enqueue(&mut op, event)?;
-
-        assert_eq!(None, next);
-        Ok(())
-    }
-
-    fn test_select_stmt(stmt: tremor_script::ast::Select) -> tremor_script::ast::Stmt {
-        let aggregates = vec![];
-        let aggregate_scratches = Some((
-            AggregateScratch::new(aggregates.clone()),
-            AggregateScratch::new(aggregates.clone()),
-        ));
-        ast::Stmt::Select(SelectStmt {
-            stmt: Box::new(stmt),
-            aggregates,
-            aggregate_scratches,
-            consts: Consts::default(),
-            locals: 0,
-            node_meta: ast::NodeMetas::new(vec![]),
-        })
-    }
-    #[test]
-    fn select_nowin_nogrp_whrt_havbad() -> Result<()> {
-        use halfbrown::hashmap;
-        let target = test_target();
-        let mut stmt_ast = test_stmt(target);
-        stmt_ast.maybe_where = Some(ImutExpr::from(ast::Literal {
-            mid: 0,
-            value: Value::from(true),
-        }));
-        stmt_ast.maybe_having = Some(ImutExpr::from(Literal {
-            mid: 0,
-            value: Value::from(hashmap! {
-                "snot".into() => "badger".into(),
-            }),
-        }));
-
-        let stmt_ast = test_select_stmt(stmt_ast);
-        let script = "fake".to_string();
-        let query = srs::Query::try_new::<Error, _>(script, |_| Ok(test_query(stmt_ast.clone())))?;
-
-        let stmt = srs::Stmt::try_new_from_query::<Error, _>(&query, |_| Ok(stmt_ast))?;
-
-        let mut op = test_select(7, stmt)?;
-        let event = test_event(0);
-
-        let next = try_enqueue(&mut op, event)?;
-
-        assert_eq!(None, next);
-        Ok(())
-    }
-
-    #[test]
-    fn tumbling_window_on_time_emit() -> Result<()> {
-        // interval = 10 seconds
-        let mut window = window::TumblingOnTime::from_stmt(
-            10 * 1_000_000_000,
-            window::Impl::DEFAULT_EMIT_EMPTY_WINDOWS,
-            window::Impl::DEFAULT_MAX_GROUPS,
-            None,
-            None,
-        );
-        let vm = literal!({
-           "h2g2" : 42,
-        })
-        .into();
-        assert_eq!(
-            WindowEvent {
-                include: false,
-                opened: true,
-                emit: false
-            },
-            window.on_event(&vm, ingest_ns(5), &None)?
-        );
-        assert_eq!(
-            WindowEvent::all_false(),
-            window.on_event(&vm, ingest_ns(10), &None)?
-        );
-        assert_eq!(
-            WindowEvent {
-                include: false,
-                opened: true,
-                emit: true
-            },
-            window.on_event(&vm, ingest_ns(15), &None)? // exactly on time
-        );
-        assert_eq!(
-            WindowEvent {
-                include: false,
-                opened: true,
-                emit: true
-            },
-            window.on_event(&vm, ingest_ns(26), &None)? // exactly on time
-        );
-        Ok(())
-    }
-
-    #[test]
-    fn tumbling_window_on_time_from_script_emit() -> Result<()> {
-        // create a WindowDecl with a custom script
-        let reg = Registry::default();
-        let aggr_reg = AggrRegistry::default();
-        let module_path = ModulePath::load();
-        let q = tremor_script::query::Query::parse(
-            &module_path,
-            "bar",
-            r#"
-            define tumbling window my_window
-            with
-                interval = 1000000000 # 1 second
-            script
-                event.timestamp
-            end;"#,
-            vec![],
-            &reg,
-            &aggr_reg,
-        )
-        .map_err(|ce| ce.error)?;
-        let window_decl = match q.query.suffix().stmts.first() {
-            Some(Stmt::WindowDecl(decl)) => decl.as_ref(),
-            other => return Err(format!("Didnt get a window decl, got: {:?}", other).into()),
-        };
-        let mut params = halfbrown::HashMap::with_capacity(1);
-        params.insert("size".to_string(), Value::from(3));
-        let interval = window_decl
-            .params
-            .get("interval")
-            .and_then(Value::as_u64)
-            .ok_or(Error::from("no interval found"))?;
-        let mut window = window::TumblingOnTime::from_stmt(
-            interval,
-            window::Impl::DEFAULT_EMIT_EMPTY_WINDOWS,
-            window::Impl::DEFAULT_MAX_GROUPS,
-            None,
-            Some(&window_decl),
-        );
-        let json1 = literal!({
-            "timestamp": 1_000_000_000
-        })
-        .into();
-        assert_eq!(
-            WindowEvent {
-                opened: true,
-                include: false,
-                emit: false
-            },
-            window.on_event(&json1, 1, &None)?
-        );
-        let json2 = literal!({
-            "timestamp": 1_999_999_999
-        })
-        .into();
-        assert_eq!(WindowEvent::all_false(), window.on_event(&json2, 2, &None)?);
-        let json3 = literal!({
-            "timestamp": 2_000_000_000
-        })
-        .into();
-        // ignoring on_tick as we have a script
-        assert_eq!(WindowEvent::all_false(), window.on_tick(2_000_000_000)?);
-        assert_eq!(
-            WindowEvent {
-                opened: true,
-                include: false,
-                emit: true
-            },
-            window.on_event(&json3, 3, &None)?
-        );
-        Ok(())
-    }
-
-    #[test]
-    fn tumbling_window_on_time_on_tick() -> Result<()> {
-        let mut window = window::TumblingOnTime::from_stmt(
-            100,
-            window::Impl::DEFAULT_EMIT_EMPTY_WINDOWS,
-            window::Impl::DEFAULT_MAX_GROUPS,
-            None,
-            None,
-        );
-        assert_eq!(
-            WindowEvent {
-                opened: true,
-                include: false,
-                emit: false
-            },
-            window.on_tick(0)?
-        );
-        assert_eq!(WindowEvent::all_false(), window.on_tick(99)?);
-        assert_eq!(
-            WindowEvent {
-                opened: true,
-                include: false,
-                emit: false // we dont emit if we had no event
-            },
-            window.on_tick(100)?
-        );
-        assert_eq!(
-            WindowEvent::all_false(),
-            window.on_event(&ValueAndMeta::default(), 101, &None)?
-        );
-        assert_eq!(WindowEvent::all_false(), window.on_tick(102)?);
-        assert_eq!(
-            WindowEvent {
-                opened: true,
-                include: false,
-                emit: true // we had an event yeah
-            },
-            window.on_tick(200)?
-        );
-        Ok(())
-    }
-
-    #[test]
-    fn tumbling_window_on_time_emit_empty_windows() -> Result<()> {
-        let mut window = window::TumblingOnTime::from_stmt(
-            100,
-            true,
-            window::Impl::DEFAULT_MAX_GROUPS,
-            None,
-            None,
-        );
-        assert_eq!(
-            WindowEvent {
-                opened: true,
-                include: false,
-                emit: false
-            },
-            window.on_tick(0)?
-        );
-        assert_eq!(WindowEvent::all_false(), window.on_tick(99)?);
-        assert_eq!(
-            WindowEvent {
-                opened: true,
-                include: false,
-                emit: true // we **DO** emit even if we had no event
-            },
-            window.on_tick(100)?
-        );
-        assert_eq!(
-            WindowEvent::all_false(),
-            window.on_event(&ValueAndMeta::default(), 101, &None)?
-        );
-        assert_eq!(WindowEvent::all_false(), window.on_tick(102)?);
-        assert_eq!(
-            WindowEvent {
-                opened: true,
-                include: false,
-                emit: true // we had an event yeah
-            },
-            window.on_tick(200)?
-        );
-
-        Ok(())
-    }
-
-    #[test]
-    fn no_window_emit() -> Result<()> {
-        let mut window = window::No::default();
-
-        let vm = literal!({
-           "h2g2" : 42,
-        })
-        .into();
-
-        assert_eq!(
-            WindowEvent::all_true(),
-            window.on_event(&vm, ingest_ns(0), &None)?
-        );
-        assert_eq!(WindowEvent::all_false(), window.on_tick(0)?);
-        assert_eq!(
-            WindowEvent::all_true(),
-            window.on_event(&vm, ingest_ns(1), &None)?
-        );
-        assert_eq!(WindowEvent::all_false(), window.on_tick(1)?);
-        Ok(())
-    }
-
-    #[test]
-    fn tumbling_window_on_number_emit() -> Result<()> {
-        let mut window =
-            window::TumblingOnNumber::from_stmt(3, window::Impl::DEFAULT_MAX_GROUPS, None, None);
-
-        let vm = literal!({
-           "h2g2" : 42,
-        })
-        .into();
-
-        // do not emit yet
-        assert_eq!(
-            WindowEvent::all_false(),
-            window.on_event(&vm, ingest_ns(0), &None)?
-        );
-        assert_eq!(WindowEvent::all_false(), window.on_tick(1_000_000_000)?);
-        // do not emit yet
-        assert_eq!(
-            WindowEvent::all_false(),
-            window.on_event(&vm, ingest_ns(1), &None)?
-        );
-        assert_eq!(WindowEvent::all_false(), window.on_tick(2_000_000_000)?);
-        // emit and open on the third event
-        assert_eq!(
-            WindowEvent::all_true(),
-            window.on_event(&vm, ingest_ns(2), &None)?
-        );
-        // no emit here, next window
-        assert_eq!(
-            WindowEvent::all_false(),
-            window.on_event(&vm, ingest_ns(3), &None)?
-        );
-
-        Ok(())
     }
 }
